+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/screen/send/integration/SendViewIntegrationTest.kt
@@ -4,16 +4,7 @@ import androidx.compose.ui.test.junit4.StateRestorationTester
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.test.filters.MediumTest
-import cash.z.ecc.android.sdk.fixture.WalletBalanceFixture
-import cash.z.ecc.android.sdk.fixture.WalletFixture
-import cash.z.ecc.android.sdk.model.MonetarySeparators
-import cash.z.ecc.android.sdk.model.Zatoshi
-import cash.z.ecc.android.sdk.model.ZcashNetwork
 import cash.z.ecc.sdk.fixture.ZecSendFixture
-import co.electriccoin.zcash.ui.common.model.TopAppBarSubTitleState
-import co.electriccoin.zcash.ui.fixture.BalanceStateFixture
-import co.electriccoin.zcash.ui.fixture.MockSynchronizer
-import co.electriccoin.zcash.ui.fixture.WalletSnapshotFixture
 import co.electriccoin.zcash.ui.screen.send.WrapSend
 import co.electriccoin.zcash.ui.screen.send.assertOnConfirmation
 import co.electriccoin.zcash.ui.screen.send.assertOnForm
@@ -21,35 +12,13 @@ import co.electriccoin.zcash.ui.screen.send.clickCreateAndSend
 import co.electriccoin.zcash.ui.screen.send.setAddress
 import co.electriccoin.zcash.ui.screen.send.setAmount
 import co.electriccoin.zcash.ui.screen.send.setMemo
-import kotlinx.coroutines.runBlocking
 import org.junit.Rule
 import org.junit.Test
-import java.util.Locale
 import kotlin.test.Ignore
 
 class SendViewIntegrationTest {
     @get:Rule
     val composeTestRule = createComposeRule()
-
-    private val wallet = WalletFixture.Alice
-    private val network = ZcashNetwork.Testnet
-    private val spendingKey =
-        runBlocking {
-            WalletFixture.Alice.getUnifiedSpendingKey(
-                seed = wallet.seedPhrase,
-                network = network
-            )
-        }
-    private val synchronizer = MockSynchronizer.new()
-    private val walletSnapshot =
-        WalletSnapshotFixture.new(
-            saplingBalance =
-                WalletBalanceFixture.new(
-                    available = Zatoshi(Zatoshi.MAX_INCLUSIVE.div(100))
-                )
-        )
-
-    private val monetarySeparators = MonetarySeparators.current(Locale.getDefault())
 
     // TODO [#1260]: Cover Send screens UI with tests
     // TODO [#1260]: https://github.com/Electric-Coin-Company/zashi-android/issues/1260
@@ -65,21 +34,12 @@ class SendViewIntegrationTest {
 
         restorationTester.setContent {
             WrapSend(
-                balanceState = BalanceStateFixture.new(),
                 sendArguments = null,
-                synchronizer = synchronizer,
-                walletSnapshot = walletSnapshot,
-                spendingKey = spendingKey,
                 goToQrScanner = {},
                 goBack = {},
                 goBalances = {},
                 goSettings = {},
                 goSendConfirmation = {},
-                isHideBalances = false,
-                onHideBalances = {},
-                hasCameraFeature = true,
-                monetarySeparators = monetarySeparators,
-                topAppBarSubTitleState = TopAppBarSubTitleState.None,
             )
         }